requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
google-genai>=1.0.0
chromadb>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
aiofiles>=23.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
